refactor: move pretalx fetching into submodule pattern
- Created src/pipeline/pretalx/ submodule following youtube/text_generation pattern
- Moved SessionRecord, SpeakerInfo, Organization models to pretalx/models.py
- Renamed fetch_pretalx.py → pretalx/fetch_records.py with fixed imports
- Updated all dependent files to import from pipeline.pretalx.models
- Added comprehensive README.md documenting usage and prerequisites
- Cleaned up shared models.py with clarifying docstring
- Updated archived files for consistency

Files affected:
- youtube/prepare_metadata.py
- text_generation/build_release_records.py
- prepare_summaries.py
- archived files

Run with: python -m src.pipeline.pretalx.fetch_records
