+++ src/pipeline/archived/2025_10_08_text_generation/generate_summaries.py
@@ -26,8 +26,8 @@ import structlog
 
 from pipeline.config import load_config
 from pipeline.logger import setup_logging
-from pipeline.models import SessionRecord
 from pipeline.paths import WorkPaths
+from pipeline.pretalx.models import SessionRecord
 
 from .models import Summary, SummaryGenerationRequest
 from .providers import ProviderFactory
