+++ src/pipeline/pretalx/__init__.py
@@ -1,0 +1,5 @@
+"""Pretalx data fetching and management pipeline.
+
+This module handles all interactions with Pretalx conference management system,
+including fetching session records, speaker information, and event metadata.
+"""
